requirements.txt
about-time==3.1.1
acme==1.1.0
alive-progress==2.3.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
bleach==5.0.0
blessings==1.6
blinker==1.4
bottle==0.12.15
Brlapi==0.7.0
cachetools==5.0.0
certbot==0.40.0
certbot-apache==0.39.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-cors-headers==3.12.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==1.16.0
docker==4.1.0
entrypoints==0.3
filelock==3.4.2
future==0.18.2
Glances==3.2.4.2
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
gpg===1.13.1-unknown
gpustat==0.6.0
grapheme==0.6.0
httplib2==0.20.4
idna==2.8
importlib-metadata==1.5.0
influxdb==5.2.0
install==1.3.5
josepy==1.2.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.1.1
matplotlib==3.1.2
mock==3.0.5
more-itertools==4.2.0
mysql-connector-python==8.0.29
mysqlclient==2.1.0
netifaces==0.10.4
numpy==1.17.4
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
packaging==20.3
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.0
ply==3.11
protobuf==3.20.1
psutil==5.5.1
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
PyDrive==1.3.1
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pysmi==0.3.2
pysnmp==4.4.6
pystache==0.5.4
python-apt==2.0.0+ubuntu0.20.4.6
python-augeas==0.5.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
rsa==4.8
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
virtualenv==20.13.1
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
